// examples/basic-react-app/src/interfaces/Reddit.ts
// Generated by https://quicktype.io

export type ErrorResponse = {
	message: string;
	error?: number;
};

export type Root = {
	kind: string;
	data: RootData;
};

export type RootData = {
	after: string;
	dist: number;
	modhash: string;
	geo_filter: undefined;
	children: Child[];
	before: undefined;
};

export type Child = {
	kind: Kind;
	data: ChildData;
};

export type ChildData = {
	approved_at_utc: undefined;
	subreddit: Subreddit;
	selftext: string;
	author_fullname: string;
	saved: boolean;
	mod_reason_title: undefined;
	gilded: number;
	clicked: boolean;
	title: string;
	link_flair_richtext: any[];
	subreddit_name_prefixed: SubredditNamePrefixed;
	hidden: boolean;
	pwls: number;
	link_flair_css_class: undefined | string;
	downs: number;
	top_awarded_type: undefined;
	hide_score: boolean;
	name: string;
	quarantine: boolean;
	link_flair_text_color: LinkFlairTextColor;
	upvote_ratio: number;
	author_flair_background_color: undefined;
	subreddit_type: SubredditType;
	ups: number;
	total_awards_received: number;
	media_embed: Gildings;
	author_flair_template_id: undefined;
	is_original_content: boolean;
	user_reports: any[];
	secure_media: undefined;
	is_reddit_media_domain: boolean;
	is_meta: boolean;
	category: undefined;
	secure_media_embed: Gildings;
	link_flair_text: undefined | string;
	can_mod_post: boolean;
	score: number;
	approved_by: undefined;
	is_created_from_ads_ui: boolean;
	author_premium: boolean;
	thumbnail: string;
	edited: boolean;
	author_flair_css_class: undefined;
	author_flair_richtext: any[];
	gildings: Gildings;
	content_categories: undefined;
	is_self: boolean;
	mod_note: undefined;
	created: number;
	link_flair_type: FlairType;
	wls: number;
	removed_by_category: undefined;
	banned_by: undefined;
	author_flair_type: FlairType;
	domain: string;
	allow_live_comments: boolean;
	selftext_html: undefined | string;
	likes: undefined;
	suggested_sort: undefined | string;
	banned_at_utc: undefined;
	view_count: undefined;
	archived: boolean;
	no_follow: boolean;
	is_crosspostable: boolean;
	pinned: boolean;
	over_18: boolean;
	all_awardings: any[];
	awarders: any[];
	media_only: boolean;
	link_flair_template_id?: string;
	can_gild: boolean;
	spoiler: boolean;
	locked: boolean;
	author_flair_text: undefined;
	treatment_tags: any[];
	visited: boolean;
	removed_by: undefined;
	num_reports: undefined;
	distinguished: undefined;
	subreddit_id: SubredditID;
	author_is_blocked: boolean;
	mod_reason_by: undefined;
	removal_reason: undefined;
	link_flair_background_color: LinkFlairBackgroundColor;
	id: string;
	is_robot_indexable: boolean;
	report_reasons: undefined;
	author: string;
	discussion_type: undefined;
	num_comments: number;
	send_replies: boolean;
	whitelist_status: WhitelistStatus;
	contest_mode: boolean;
	mod_reports: any[];
	author_patreon_flair: boolean;
	author_flair_text_color: undefined;
	permalink: string;
	parent_whitelist_status: WhitelistStatus;
	stickied: boolean;
	url: string;
	subreddit_subscribers: number;
	created_utc: number;
	num_crossposts: number;
	media: undefined;
	is_video: boolean;
	url_overridden_by_dest?: string;
};

export enum FlairType {
	Text = 'text',
}

export type Gildings = Record<string, unknown>;

export enum LinkFlairBackgroundColor {
	D3D6Da = '#d3d6da',
	Empty = '',
}

export enum LinkFlairTextColor {
	Dark = 'dark',
}

export enum WhitelistStatus {
	AllAds = 'all_ads',
}

export enum Subreddit {
	Javascript = 'javascript',
}

export enum SubredditID {
	T52Qh30 = 't5_2qh30',
}

export enum SubredditNamePrefixed {
	RJavascript = 'r/javascript',
}

export enum SubredditType {
	Public = 'public',
}

export enum Kind {
	T3 = 't3',
}